requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.16.0
scrapy==2.11.0

# Backend API
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pymongo==4.6.1

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
pandas==2.1.4
python-dateutil==2.8.2

# Email notifications
sendgrid==6.11.0
